This commit signals the begining of the waterfall model.
It also changed the code in such way to compile with clang.

Also on :substitute --interactive, add a [cC]ontinue to next line choise to
the interactive dialog.

SLRE: add a (int) .total_caps to the info structure to set the exact number of
the captured substrings.

Re.exec(): catch the RE_CAPS_ARRAY_TOO_SMALL_ERROR and reallocate array.

To document also, that 'pattern` can be set to ignore case, same as pcre does:

  (?i)...