Change the str_dup() to accept a len argument and avoid a recomputation. In all the cases except one (which is called at most a couple of times):   - the size is already known   - they called usually on undo actions! that is in almost every change!
First: This wouldn't be possible if the standard strdup() used instead.

Now, this is a free of charge (and looks a big one actually) optimization
since avoids a strlen(), in one of the most used functions and every bit
counts in C;  otherwise there is no reason to torture yourself, when there
are uncountable ways today to abstract the details, either by using another
programming language, or even to use one of those C libraries that offer a
higher interface.
I think that is the reason why there is no such universal library, though
there are quite a few that permit crazy smart stuff, but with even a bit
of extra bits of overhead.

Yes, i think, every bit counts in C, do not lie to yourself.