on [yY]ank: send yanked text to the xclipboard, if $DISPLAY is not NULL and if xclip (for now) is available.
Also:
Simplify register code.

Add a __string_new_with() function, that size of the string is given as an argument.
Again this is an (free and without a risk) optimization. This function is considered
as a low level function, that expects sanity, as it doesn't do any kind of check for
valid conditions. This is because, any mistake will be caught before in the toolchain,
as the data is produced by own code and therefore is assumed valid.