Add a first working draft of a `search on file for pattern` feature.
The command it takes a pattern and at least a filename as arguments:
  :vgrep --pat file[s]

This opens a unique window intented for searches and accessible actually with:
  :searches  (though it might be wise a `:copen' alias (to match vim's expectation))

This window should open a frame at the bottom, with the results (if any) and it will
set the pointer to the first item from the sorted and unique list.

'\r' should open the filename at the specific line number at the frame 0.

No other feature.

Now, this was the last big feature that left unimplemented, besides process
managment which might be better if could be handled in an outer scope, though
for now there is no sense of an outer space; perhaps a couple of bits are justified.