This is to document a macro side effect that it might gets unoticed, see:
now, what happens here:
  if (IS_DIGIT (*sp++)) ...

is that the expression (*sp++) is evaluated twice, so the pointer is incremented twice,
so it might not be happen, what the poor programmer's mind thought that it will happen.

Also, on normal mode: CTRL_[ax], works also for octal and hex numbers. Still it can not
operate on binaries, though it operates on all the other bases and also for chars;
even for multibyte characters (so an 'α' becomes 'β' on CTRL_A (though for now is limited
to known ranges and should be extended with carefull research))

Added special registers:
 - % expands to current absolute pathname
 - / expands to the last search
 - : expands to the last entered command
 - _ blackhole register, which stores and puts nothing

It should mentioned the +* (plus and star registers from previous commit) for communication
in X.