Add :diff[buf] commands.
This feature requires (for now) the `diff' utility.

The :diff command open a dedicated "diff" buffer, with the results (if any) of
the differences (in a unified format), between the buffer in memory with the one
that is written on the disk. This buffer can be quickly closed with 'q' as in a
pager (likewise for the other special buffers, like the message buffer).
Note that it first clears the previous diff.

The :diffbuf command gives the focus to this same buffer.

Another usage of this feature is when quiting normally (without forcing).
In that case a dialog (below) presents some options:

  "[bufname] has been modified since last change
   continue writing? [yY|nN], [cC]ansel, unified [d]iff?"

 on 'y': write the buffer and continue
 on 'n': continue without writing
 on 'c': cansel operation at this point (some buffers might be closed already)
 on 'd': print to the stdout the unified diff and redo the question (note that
         when printing to the stdout, the previous terminal state is restored;
         any key can bring back the focus)

Also added a find_executable_in_path (char *name) which returns the first found
`name' in $PATH and (0 == access (/path/name, F_OK|X_OK))

Add a development tracker file called DevDo.md.