Fixes and new semantics for the history completion mechanism.  - the ARROW_UP key starts from the last entry set in history, and scrolls down to the    past entries
 - the ARROW_DOWN key starts from the first entry, and scrolls up to most recent

Now matching current line with past entries should work more properly.

This also fixes some strange combinations with xclipboard (both selections) registers,
that returned the text in reversed order.

Also before those registers were set by default as LINEWISE, and it was impossible to
use them in insert mode or in the command line (on CTRL('r')).

This is the first bug fix commit. Two futures remain to implement to the library,
of course many enhancements and uncountable conditions, but (hopefully) this signals
the start of the stabilization period.