Added a mechanism for actions on current word, which is triggered with 'W' on normal mode (not many choises for now (though it's easy to add existed ones and ones that make sense like `delete' or `yank'), just enough for the demostration)
Replace str_tok() with str_chop (), with the same signature except the delimiter
argument, which for this function is a char.
This allows to avoid an extra copy of the string passed as argument to strtok(),
since this function modifies the string.
Also (and since a newline ('\n') was the delimiter in all the calls to str_tok()),
the operation should be quicker now, as it scans the string for a matched byte and
not for a combination of bytes.

This also fixes a subtle bug (a pointer incrementation that goes out of bounds),
and which was present from the begining and caught by valgrind.
And for second time: How the C developers coped with memory leaks and similar in the
pre valgrind days?