This commit is maybe the first without a new big feature, and as usual includes quite many bug fixes and tiny enhancements (that in the previous log messages never mentioned). But the real reasons for this commit are the two explained below.
The first is the following comment in one of those code blocks.

  /* This code block reflects perfectly the development philosophy,
   * as if you analyze it, it catches two conditions, with the first
   * to be a quite common scenario, while the second it never happens
   * at least in (my) usual workflow. Now. I know there are quite few
   * more to catch, but unless someone generiously offers the code i'm
   * not going to spend the energy on them and can be safely considered
   * as undefined or unpredictable behavior.
   */

The second is that this ends up the zero cycle, after nine months of development,
and probably is a good idea to tag it with a way in git.

The trueth is that there a couple more small practical features that needs implementation,
but the library proved in the last development that it can be usefull as an underlying
machine, as it was destined to do from the begining. There is a missing important
piece still, but this should happen outside of the library.

As a recap.
But whatever the plans and the ideas, this is now a quite usable indepented editor,
that runs the operations in full speed by using quite few of resources. In a small
test (opening the same file), this requires the fewer bits of memory of all the others
editors, besides plain ed-like editors, with the jed editor from my friend John, being
the second. And it should be mentioned here that the desire was to be usable when editing
C sources, as if you check there is no other filetype, as my time spending around C.

As such, i'm happy, though the real happiness can be found elsewhere and not in coding.