This is an ought to do commit, that handles tabs or characters that occupy more than one cell. The last, achieved by using the following wcwidth() implementation:
/*
 * https://github.com/termux/wcwidth
 * The MIT License (MIT)
 * Copyright (c) 2016 Fredrik Fornwall <fredrik@fornwall.net>
 *
 * This license applies to parts originating from the
 * https://github.com/jquast/wcwidth repository:
 *
 * The MIT License (MIT)
 *
 * Copyright (c) 2014 Jeff Quast <contact@jeffquast.com>
 */

Also we should mention here Markus Kuhn that worked for years in UTF8 issues,
to standardize it and finally gain universal adoption.

   Many Thanks.

Note that there is a certainity that the code doesn't handle all the possible conditions
that can be met, but in any case this is a huge step forwards, and because of this,
the editor can be used in broader environments, other than C programming which was att
the begining one of its purposes.