Ongoing and incomplete work on syntax and file types. Add file types for sh and make, which intented for testing and development.
So, this is just to save state and probably and for first time, we'll see some
likewise on the next commits.
But anyway this might contains bugs. That doesn't mean of course that the other commits
don't. But i'm not certain (trueth is that never was and my opinion is that never we
can) for the future (needless and unrelated stuff anyway - ignore).