User Interface (ui): Normal mode: (first draft) add CTRL('o') and CTRL('i') to jump to previous locations in the jump list.  This is like scrolling to (back|for)wards to the history, so the semantics differ from vim.
Normal mode: on 'g' pressing 'b', opens the link under the cursor, in a browser.
This requires the elinks text browser to be installed. It uses the -remote option
to open the link in a new tab, so elinks should be running, otherwise does nothing.
Implemented as an extension to the usr.c unit and serves as an example and as a
test to the mechanism.

Inner Development: a couple of buf fixes but mostly reorganization, so this might
looks like a big commit, but it isn't.