Delete the needless "usr.h" and remove any inner dependency between the local and user namespace.
Though is not used in current code, make more robust the menu selection/completion
system on various actions, by passing an extra argument to the callback functions,
which is simply the selected action. So far, only the first char of the selected
action was passed, to indicate the selected action and this what is actually is
being used now.

Fix a memory leak in __u_cw_mode_cb__(), by free'ing the iterator.