Preliminary introduction of a math expression facility. It is limited supported in (char|line)wise visual mode. The result is printed in the message line and stored to the 'M' register. In case of an error the focus move to the message buffer, hopefully with a detailed error. Enabled via HAS_EXPR=1.
The underlying code is from tinyexpr project:

https://github.com/codeplea/tinyexpr

zlib License
Copyright (C) 2015, 2016 Lewis Van Winkle

Many Thanks for the compact and quite expressive code.

But this introduces also and challenges, which going past the bounds
and the scope of this project and with uncertain route.

(note that tcc can not compile tinyexpr, only if we discard the `const' qualifier
of quite of many definitions - so the future is even more uncertain after that)

BTW: its been a year since initialization. This gets weird (all these time related
daily moments lately, that syncronized and connected with each other).