Expose tmpfname_[new|free]() through the File structure.
Introduce a qualifiers mechanism (kinda like S_Lang's qualifiers,
but in C they should be declared as function arguments (though now
that i'm thinking, perhaps they can be implement through variable
length arguments, or with a stack mechanism (both with a price of
course)), by using designated initializers.
(it can been seen also as a "named arguments mechanism", as while
the macro sets the defaults,the user can still overide them during
calling, through anonymous structs)

Idea from: https://github.com/mcinglis/c-style

Added -Wno-override-init in the CFLAGS as it is included in -Wextra
and the compiler warns for those designated initializers (probably for
side effects).

Change the signature for win_buf_new() to use the above mechanism.

Note that, quite many of the macros, designed to make clear to the
reader's eye, the meanings of the called function arguments.
So this mechanism is best for this and allows extensibility, without
to be obligated to add new arguments to the function when extra logic
is required.

Change some common function names in tinyexpr, as easily can class
with others (they did in a private experiment with another project).
(sorry this is a side effect of the building system; i'm #include'ing
files instead to proper use the Makefile. This is to simplify things
while building, but mostly should be the desire of a sane dependency
mechanism in C (no complain..., just probably a secret desire or even
a reaction of the lack of it). Again, no complain. I just need to work
more)