Introduce a slightly modified argparse to handle command line arguments. See at:
https://github.com/cofyc/argparse

The MIT License (MIT)

Copyright (c) 2012-2013 Yecheng Fu <cofyc.jackson@gmail.com>

Many thanks.

Options:
    +, --line-nr=<int>    start at line number
    --column=<int>        set pointer at column
    --ftype=<str>         set the file type

The modification to argparse was made specific to support the + option,
which is a standard to unix editors (both vim and jed and probably others
support this). This new option is SHORT_OPT_HAS_NO_DASH and can be set
as a flag.