Use the example code from man printf(3), to calculate the size for variable length string type function arguments. When first tried this code older compilers complained on -Werror=alloc-size-larger-than= or -fsanitize=undefined with:
  argument 1 range [18446744071562067968, 18446744073709551615]
  exceeds maximum object size 9223372036854775807
  in a call to built-in allocation function '__builtin_alloca_with_align'

current code compiled with clang-9.0.0 (with -fsanitize=undefined) and with
gcc-9.2.0 without above flag, but both with -Werror.

Before, this size was set as MAXLEN_LINE * 2 and that was not reliable,
though the code never asked for more than this limit.

Also, remove -Werror when $CC is tcc, which warns when compiles tinyexpr with:

  assigment discards qualifiers from pointer target type

and it seems that -Wno-error=discarded-qualifiers doesn't have effect

Modify a bit argparse_help_cb() to return a value and avoid an exit (again for
tcc, that warns), and handle the returned value.