This is to save state as easily can go out of control after a couple of intrusive changes.
This introduces the root class that hold and manipulates editor instances, and as
a first result, it also should handle gracefully now those instances.

This made also possible to hide all the exposed data to the private environment,
which is now only accesible through [sg]etters (with a consistent way for all and
without exceptions (as before)).

Introduces to the user space the public variable __E__ which is a pointer to this
root class instance. This class should have all the necessary information for the
library to be fully functional. As it is early to judge and still there things that
need to be handled, it compiles without warnings and runs without memory leaks.

Also enhance @info command with details about the current editor, current buffer and
current window (with their own command line switch; without an argument this prints
details about current buffer).