This finalizes the underlying structures and interface.
A couple of details might change in future (mainly visibility rules) but this is it.

Added :ednew|ednext|edprev|edprevfocused commands to create and cycle through editor
instances, which are totally independent to it's other.

The root class holds the information for those instances and creates a prototype that
is being used to instantiate the structures and a couple of properties.
All the editor new instances inherit from this prototype, and as well to their children.

This development permited much simpification for the user of the library and set up
roles and semantics. It also provides ways to overide functionality, though for now
any such change will have global effect and possible this is not look so wise.