Introduce an 'autosave' buffer option, that is checked when leaving insert mode, and if the buffer has been modified and the interval time is > autosave, then the buffer is written to the disk.
It can be enabled at runtime with
  - :set --autosave=[int]
  - on invocation with the --autosave=[int] command line option

This value is meaning minutes, that are translated internally in seconds and
can not be larger than a day. A value of zero deactivates functionality, and
this is the default.

This uses clock_gettime() and by default CLOCK_REALTIME.

Also, expose many internal functions to their corresponding structures.