Added a mechanism to check for an unbalanced pair of objects, like "()[]{}", which are the default. It can be used from the command line with :@balanced_check [--range=], if `range' is ommited, default the whole unit. This is a violation (for now) of the rules, though the default (current line), doesn't make much sence here.
Also it can be used in visual linewise mode through tab completion and the popup
menu.

Note that the code doesn't handle multiline comments. But seems to handle
singleline comments and also continues, if one of the chars is on a string,
or is included in single quotes.

If finds an unbalanced pair, it redirects a proper message to the message buffer
and then opens that buffer, otherwise a simple message shows in the message line.

Note that it passes the main development file which is almost 15000 lines of code.