Add a backup functionality that could be set during runtime or on initial invocation.
Relative command line switches:
  --backupfile
  --backup-suffix=[str] (default: ~)

Likewise for the set command.

This generates a backup filename as:
  dirname (filename) + / + .filename + backup-suffix

Also a @bufbackup command is provided for convenience. The set command
should be used first to set the backupfile option (if it is not already
set).

A :set --no-backupfile option is also provided for the inverse.