Initial development of "This", which in this context means the application.
The public pointer variable that controls the application, and which has all the
necessary information, named __THIS__.

The public pointer variable that holds the methods of this class, named __SELF__
(though we could do it without, but i'm not sure when the cast'ing taking place,
to avoid an overhead).

Introduce a tiny indirection (the first one in the whole application/library),
to always pass the current Root Editor instance as a first argument to the calling
functions, by using __THIS__->__E__ (which is already exposed in the library).

The "This" and "E" macros are provided for syntactic sugar/convenience, with
the first used for the application methods, and the latter for calling the
methods, from the Root Class of the library.

Finally and this is the most interesting future in this commit, introduce a mechanism
to direct the machine to run a command at the startup, and with the possibility to
exit afterwards. This will allow to create utilities, like a sed functionality with
the following way:

vedas --ex-com="s%::--pat=some_pattern::--sub=replaced::--interactive" --exit file

The "::" is being used as a delimeter, so that the arg parser will not be confused
by the whitespace in between.

And lastly remove the $myed ugly macro and lower the overhead in many occasions at
the same time.