Early development of a scripting interface.
This is based on Tinyscript at: https://github.com/totalspectrum

Copyright 2016 Total Spectrum Software Inc.

 +--------------------------------------------------------------------
 ¦  TERMS OF USE: MIT License
 +--------------------------------------------------------------------

For an example use:

   vedas --load-file=lib/i/example.i

It is a really really tiny scripting interface (with just little bit over 1000
lines of code), but with enough features to load and evaluate very simple code.

Many many Thanks to the Tinyscript project.