Add a shared register functionality.
It can be used from all the editor instances, to share data via this register which it
is accessed for both reading and writing by all.

It can be used normally as any other register, plus via a menu item in visual linewise
mode. The default char for this is "`".