Replace memmove(), memcpy() and strrchr() with internal functions.
Clean up also some noise while compiling.

Fix a bit the unreadable README.

Also this is the end of a series of commits, that integrated code that has been
tested for a long in outer namespaces. The status of the code might be in a bit
unstable state for sometime, as too much code integrated in a quite short period.

But all the major features have been implemented (except one).

So this signals the beginning of a stabilization period.

As a reminder this is the result of twenty months of (mainly intensive) coding.