Extend the :diff command to accept "--origin" as an argument.
In that case if "backupfile" was set, the command compares the contents of the
"backupfile" with the current buffer. Otherwise it issues a warning.

Move Imap Class to the core, as it is a datatype.

And as a constant progress from now, make function names consistent with their role
and their class name. Likewise for Makefile variables, macro names, variable names.

Bump REV[ision] to 2 and remove VERS[ion]. That means a naming schema libved-0.2.so
for the library and veda-02 for the executable. The zero here denotes the API which
will become 1 when it will be stabilized (which is not that far in fact, as the API
is quite stable at least as a mechanism; the only thing that it can change is function
signatures).