Remove from the extended namespace two global variables.
Now what is left is just one (__This__), that holds all the required information.
The THIS macro refers to that variable.
Likewise the This macro refers to the methods.

Also, do not use My(Class) anymore in the core code, and make it consistent with
the outter namespace, that uses directly a class, for example:

   String.method (...)

Note, like the My macro, this needs one of the buf/win/ed structures to be declared
as this.

To our Little Blue Angel.

This is dedicated to our little brother (26 years old) Βασίλη, a passionate fighter
for the justice. And justice is what he might expect and hopefully he already works
from where he is now (we don't know where but he is somewhere now for sure (probably
nobody really dies ever and this quite logical if you think about)).