Do not try to execute a record if there is nothing already recorded.
All this work, is dedicated to Zoi, for thirty years a gorgeous companion and mother
of our four kids as Zoi means life here actually.

She was born a day like this, when the humans touched the moon, but a day like this,
it was also the first whole day i remember crystal clear in my life (20 July of 1974).
I remember walking with my mother coming from my beloved grand parents. My father had
to go to the army that same day, because there was a war. I remember an extraordinary
bright night sky, though the moon was new.