Set the current buffer when changing focus to other Editor instances (mainly to setup current directory, which before miscalculated).
Fix a segfault on buffer deletion. This code used to work before as a direct call to
the actual function, but after the recent changes (where we use the abstraction layer),
the "this" variable ended up as a NULL reference (so naturally the macro failed), and
hence the segfault.

There might be other cases like this one.