Introduce a mechanism to automatically save an image of the current process.
:set --save-image=[1|0] to [en|dis]able the functionality.

It can be reloaded with --load-file=[image_name] at the invocation.

A bit more detailed information will have to wait the stabilization.

Introduce path_basename_sans_extname (char * name) that returns the basename of
`name' sans the extension name. It returns an allocated string. Perhaps a shared
string to the Path class is justified. Otherwise it should be provided a narrower
functionality to returns the first idx or|and the last idx or|and the len match,
or it could be done with a regexp. In the last case, Path should be connected with
other Structures, plus one Re instance should be easily available.

By the way! An integrated interpreter, with sooo cheap and straight access
to the whole system, is fascinating and opens a big box of possibilities.