Allow to exit after an early file evaluation (using --load-file=`file switch).
This can be used at least in this state, as a testing mechanism for
the interpreter - actually it acts like an interpreter; as in this early face,
it still nothing has been loaded, except the library machine, so the
interpreter can use it fully as it likes.

So and as a intented documented behavior, that is what it happens
with the --exit command line switch following a --load-file=`file'.