Add :set --persistent-layout.
The purpose for this is to enable a persistent behavior while loading and saving
editor layout.

With this option an accidentally :q (without global) and if there is more than one
editor instance (usually that is the case with an extensive layout (projects)), it
changes the current layout (that the intention is to keep it). This is because the
current editor instance that is "quit-ed" it won't be recorded. With this option
this is not happens, though it might miss some information, like the current pointer
position. There is a way to even keep this information, but code has to be written
to extend a bit the logic and with a minimal overhead.