Prevent from some very often video allocations and free's and dedicate one such type that can be used for these kind of operations.
Put a couple of guards, to prevent from segfaults and exit a little bit more gracefully,
when the available LINES and|or COLUMNS, are less than the required to be functional.
This is on initialization and on SIGWINCH.

Also fix the sigwinch handler from an eternal loop (this is because of a code change,
that got unoticed, because of the simple fact, that it never occurs in my environment
a SIGWINCH signal). The constant environment for more than fifteen years is full screen
terminals, that spawned in a self made tiny window manager that maximizes the created
windows to get the whole precious screen space, and which is controled by using only
the keyboard for all the operations.

The first WM of that kind and the root of all is the legendary ratpoison, so this is
dedicated to those wise made minimalistic environments, that handle so efficiently and
with so respect the available computer resources, by being at the same time of course
and the faster environment of that kind. Many many thanks to all those people, who put
effort to develop such environments, that gave the chance to the poor human being, to
feel that he can as well interact with those cold machines, with a sensible peacefull
and simplistic way, and without over and fancy complication.
 he with a sens, and not the opposite.
, but especially to those who are using them, as
they find them sensible.