Add an --exit-quick command line argument. The intention is to exit all the running instances, when quiting the current buffer.
Add a __man shell script, that will display the first given argument as a man page,
if exists. A 'q' from normal mode, or an :q in command line, will quit the editor,
without confirmation.

Alt with a combination (like the right key), produces an ESC-ESC-[(capital), and in
normal mode a C or a D, can even change or delete text. Try to catch that by ignoring
the alt key.