Expose functionality.
Now, a buffer can exit in a paused state, that can be resumed by re-calling the main function.

Add an option to the readline interface, that allows to return after a tab completion.
Note that the library has been verified that can be used as a readline mechanism. This can be done, by
disabling the default commands, enter directly in command line mode, and at the end
put the buffer in paused state. A buffer can disable the statusline, and the "~" on empty lines,
the window can have a custom topline and there is no posibility to enter in normal mode, so
it is just the underline machine, without any other functionality.
This addition assists to accept fast a menu item, that is triggered by the tab completion.

Testing: Call the readline callback functions early, instead at the end.

Do not change the returned value as it has been set in the code in ed_loop (). Previously this was zero.

Both above changes could have some consequences and they could break expectations.