Include the tinyexpr C unit, when building the extended library.
Add a (void *) "user_object" to the readline structure, so that it can
be used by the user. The (void *) "object" field remains for internal use.
Add the corresponded [sg]etter methods.

Now, the [count] (goto line number function) and if it is larger than
the buffer lines, gets as value the 'number' of buffer lines.
Previously this function returned in that condition.

Interactively the old behavior remains, as the operation it is aborted
early when `count' is larger than `num' lines.
As the new behavior it is like the eof (end of file) function, a 'G' can be used
in normal mode to get this.