Add a note about the current development status.
It's been almost two years since the initialization, and though I might try to
backport the development that happens elsewhere, this might be the last commit.

Because of that, this work, is dedicated to those who lost their battle with an
unjustified way, fighting for a common sense, justice and against illusions and
ignorance. A yound friend said once a little before we loose him, so soon:

"Even if We will loose the battles, We will keep and keep fighting"

Probably, this unbelievable belief is the only one that can lead and guide us
to our journey, no matter how paradox it might looks like in this vanity.